refactor: 에러도 별도 애그리게이트 에러로 분리
