+++ src/main/java/com/techfork/domain/activity/bookmark/exception/BookmarkErrorCode.java
@@ -1,0 +1,25 @@
+package com.techfork.domain.activity.bookmark.exception;
+
+import com.techfork.global.common.code.BaseCode;
+import com.techfork.global.response.ReasonDTO;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+
+@RequiredArgsConstructor
+public enum BookmarkErrorCode implements BaseCode {
+    BOOKMARK_NOT_FOUND(HttpStatus.NOT_FOUND, "BOOKMARK404_1", "북마크를 찾을 수 없습니다."),
+    BOOKMARK_ALREADY_EXISTS(HttpStatus.CONFLICT, "BOOKMARK409_1", "이미 북마크한 게시글입니다.");
+
+    private final HttpStatus httpStatus;
+    private final String code;
+    private final String message;
+
+    @Override
+    public ReasonDTO getReason() {
+        return ReasonDTO.builder()
+                .httpStatus(httpStatus)
+                .code(code)
+                .message(message)
+                .build();
+    }
+}
